remove out files
