+++ README.md
@@ -11,6 +11,13 @@ Bank Application implemented in C++. Features include:
 
 Runnable from most C++ compilers such as [programiz.com](https://www.programiz.com/cpp-programming/online-compiler/)
 
+Mac Terminal Commands
+```
+// Navigate to project folder
+g++ -std=c++11 main.cpp BankAccount.cpp -o atm.out
+./atm.out
+```
+
 ## Note
 
 Initial run will create BankAccount object with default values. Select 5 from the options menu to obtain account number and utilize default PIN to play with other features or create a new account by selecting option 1.
